+++ lightning/src/ln/mod.rs
@@ -61,6 +61,9 @@ mod onion_route_tests;
 #[cfg(test)]
 #[allow(unused_mut)]
 mod monitor_tests;
+#[cfg(test)]
+#[allow(unused_mut)]
+mod shutdown_tests;
 
 pub use self::peer_channel_encryptor::LN_MAX_MSG_LEN;
